reorganize folder structure; reorganize module 'Validation'
